Fixed bug with uniform randomization in which if you had a variable with more than 250 values:
task.randVars.uniform.color=1:360;

it would only give you the randomization of the first 250 of them. This was because there is a default block length to calculate for which was (arbitrarily set to 250 - to be large). Now, the default block length is set to 250 or the length of the parameter which ever is greater, so this should be now fixed