This changeset fixes a problem that cropped up with matlab2013a. Essentially the old system call to go full screen in CGL no longer worked (it was deprecated in 10.6) so put in some switch code to use the new call for matlab2013a. It turns out the new call is also deprecated in 10.8 (go figure) so this may have to be updated again.
Also added code to go full screen from cocoa (which may be a work around for the CGL). It needed to have a uwait in the close to make it stable, but tested opening and closing many times and it seems not to crash, so that seems to have fixed it. This functionality is accessed by setting the param mglSetParam('useCGL',0).